LayoutTests/imported/w3c/web-platform-tests/css/css-page/cssom/margin-002-expected.txt

CONSOLE MESSAGE: TypeError: undefined is not an object (evaluating 'marginRule.style')

Harness Error (FAIL), message = TypeError: undefined is not an object (evaluating 'marginRule.style')
